Module cryptoxide::ed25519

ED25519 Signature Scheme

use self::cryptoxide::ed25519;

let message = "messages".as_bytes();
let seed = [0u8;32]; // seed only for example !
let (secret, public) = ed25519::keypair(&seed[..]);
let signature = ed25519::signature(message, &secret[..]);
ed25519::verify(message, &public[..], &signature[..]);

Constants

PRIVATE_KEY_LENGTH

PUBLIC_KEY_LENGTH

SEED_LENGTH

SIGNATURE_LENGTH

Functions

exchange
Curve25519 DH (Diffie Hellman) between a curve25519 public key and a ed25519 private key

keypair
Create a keypair of secret key and public key

signature
Generate a signature for the given message using a normal ED25519 secret key

signature_extended
Generate a signature for the given message using an extended ED25519 secret key

to_public
generate the public key associated with an extended secret key

verify
Verify that a signature is valid for a given message for an associated public key